Remove Tensor::ConvertToHostTensor() overload.
We are rethinking the approach for copying tensors/buffers across devices. One
issue with the current design of Tensor::ConvertToHostTensor() is that it is
not implementable for DenseGpuTensor, as it requires access to GPU stream for
the conversion.

In this CL, we remove the Tensor::ConvertToHostTensor(Location, uint32_t)
overload, as we will refactor the HostContext and error handling functionality
out of Location into a new concept, ExecutionContext, to be introduced in a
follow up CL.

PiperOrigin-RevId: 308364950