[TFRT:Data] MapDataset should enqueue work after the input value is available.
Currently MapDataset enqueues a task before the input value is available. And the task uses input_value.AndThen(...) to run map function after the input becomes available. If the input_value is computed by a thread in the blocking threadpool (e.g. it comes from TFRecordDataset) and it is available after we the enqueue this task, then it is possible for a thread in the blocking threadpool to run the map function, which is against the TFRT threadpool usage guidance.

PiperOrigin-RevId: 310441010