Add MemoryDataset which can cache and repeat another dataset.
This CL added MemoryDataset which wraps around another Dataset instance and repeats the dataset for arbitrary number of times. The data obtained from the underlying dataset will be put in an in-memory buffer until all data in the underlying dataset are enumerated. After that it just enumerates data in the in-memory buffer.

imagenet_preprocess.benchmarks.mlir is updated to use MemoryDataset to minimize the impact of I/O latency on the benchmark results.

PiperOrigin-RevId: 311263621