[TFRT:Data] Add FilterDataset that can handle asynchronous EOF.
FilterDataset takes a user-defined function. The user-defined function takes value from the underlying iterator and returns a boolean value. FilterDatasetIterator forwards the value from the underlying iterator to the GetNext(...) caller iff the user-defined function returns true when applied to the value.

FilterDatasetIterator::GetNext(...) is non-blocking and it ensures in-order delivery.
PiperOrigin-RevId: 312374130