Add Value type for TFRT interpreter.
Value is a type-erased data type for representing synchronous values. It will
be used for defining synchronous kernels and used in TFRT interpreter.  This is
the counterpart of AsyncValue for defining asynchronous kernels.

The Value class has an inplace storage that avoids heap allocations for small
object types. Currently, we set the inplace storage size to 64 bytes to avoid
heap allocations for storing DenseHostTensors which takes 48 bytes while
keeping the Value to be the size of a cache line (64 bytes).

The Value class can store any type regardless of their move/copy-ability. The
memory layout for different payload types is as follows:
* For moveable types with size <= 56 bytes, the payload is stored in place
  with no heap allocation.
* For moveable types with size > 56 bytes, the payload is stored out of place
  with a heap allocation.
* For non-moveable types regardless of size, the payload is stored out of
  place with a heap allocation.

TODO: Add the conversion between AsyncValue and Value to allow for the
interoperation of TFRT interpreter and executor.
PiperOrigin-RevId: 318915963