Add support for getting a base class reference for a stored derived class in Value.
Unlike std::any, Value supports getting a reference to the base class of the
stored type. For example, the following code works:

struct BaseClass {};
struct DerivedClass : BaseClass {};

Value v{DerivedClass()};
auto& base = v.get<BaseClass>();

This feature is useful for defining synchronous kernels that take a base class
as an argument, e.g.

// Tensor is a base class. `t` to refer to any subclass of Tensor.
void PrintTensor(const Tensor& t) {
  tfrt::outs() << t;
}

One restriction to this capability is that either both the base class and
derived class are polymorphic or neither are. The reason for this restriction
is that we do not need to adjust the pointer for casting from a derived class
to a base class when both the base and derived class are polymorphic or
neither are. Lifting this restriction would require us to assume the memory
layout of class types, which would make the code non-portable.

TODO: Add type assertion to prevent getting a base class reference for the
stored type in the presense of multiple inheritance or virtual inheritance.
This will make the debug type checking more robust.
PiperOrigin-RevId: 321436963