[TFRT:Data] Optimize InterleaveDataset by reducing number of lock grabs needed by its control flow logic
Prior to this CL, the InterleaveDataset will grab lock once to update one value in the output buffer. After this CL, the InterleaveDataset will grab lock once to update a batch of values in the output buffer. This change reduces the lock contention between the thread calling GetNext() and its token owner thread.

PiperOrigin-RevId: 325925952