[TFRT:Data] Optimize io::PrefetchingIterator's fast path by allowing its GetNext() to return value from buffer directly
Prior to this CL, if the blocking threadpool task is running when the GetNext() is called, the caller will construct and enqueue a pending IterationResult into the output_buffer. And this pending IterationResult will be materialized later by the blocking threadpool task.

The previous approach has the drawback that the blocking threadpool task might do non-trivial control flow logic when it materializes the output value due to the AndThen(...) logic. This complicates the performance debugging of the data pipeline execution. And it also incurs additional performance overhead due to the use of IndirectAsyncValues etc.

This CL simplifies the control flow logic and improves the performance in the common case, by having GetNext() return a value from the prefetch_buffer directly when possible.

PiperOrigin-RevId: 327144207