[TFRT:Data] Update PrefetchDataset to prefetch exactly the user-specified number of values
Prior to this CL, if prefetch_num is N, PrefetchDatasetIterator::GetNext() will fill the internal buffer with N-1 values and return 1 value to the caller. This CL fixed this issue so that the iterator will fill the internal buffer with N values.

The CL also added LogDataset to make it possible to test the internal behavior of PrefetchDataset and possibly other datasets in the future.

PiperOrigin-RevId: 329346949