[TFRT:Data] Update MapDataset to execute the user-provided function inline
Prior to this CL, MapDataset always enqueues the user-provided function to the non-blocking threadpool. The downside of this approach is that it could be slower than executing the function inline if the function itself is cheap (e.g. less than 200 ns). Note that the overhead of running enqueueing a function to the threadpool is about 200 ns.

This CL updates the MapDataset to always execute the user-provided function inline. And per TFRT kernel design principle, we expect kernels themselves to offload expensive operations to threadpools. This allows us to achieve the best of both world, i.e. a kernel is executed in the threadpool if and only if the kernel is expensive.

This CL also updated a few image-related and proto-related kernels to follow this TFRT kernel design principle by having them enqueue expensive operations to the threadpool.

PiperOrigin-RevId: 329624042