Initial CL to move async dispatch functions out of HostContext as global functions and add ExecutionContext as an argument for the async dispatch functions.
This is in preparation for passing priorities from RequestContext/ExecutionContext to CWQ.

In this CL, the global async dispatch functions simply forward the call to the implementation in HostContext. In the follow up CLs, we will remove the implementation in HostContext and change all the call sites to call the global functions intead.

PiperOrigin-RevId: 331228725