Use task names to replace HostIds in distributed runtime.
There are currently two ways to identify a task: the task name (e.g., "/job:.../task:...") and the host id, both can be used to identify a host in the cluster. The task name is more general in associating with remote devices, being friendly to cluster configuration changes, and being compatible with current TensorFlow. In addition, the followup cluster initialization change will make the cluster configuration to be set remotely and updated dynamically, which will break the HostId which essentially assumes a consistent offset of the task-address in a cluster config array to represent a task.

To avoid the burden of keeping both concepts consistent across the cluster, we replace HostIds with TaskHandles, which are opaque handles which uniquely represent a task in the cluster. Different from HostId, the TaskHandle is only meaningful inside the host that generates it, and cannot be serialized across different hosts. Also introducing a helper get_task_handle kernel to get the task handle from its name.

Example:
  %remote_task_name = "tfrt_test.get_string"() { value = "/job:worker/task:1" } : () -> !tfrt.string
  %task_handle = tfrt_dist.get_task_handle %client_context, %remote_task_name
  %c0 = tfrt_dist.remote_register %client_context, %task_handle, ...
  %c1 = tfrt_dist.remote_execute %client_context, %task_handle, ...
PiperOrigin-RevId: 341514149