Fix memory leak bug in HostBuffer
When the memory allocated by HostBuffer is > align boundary, it
initializes HostBuffer with the incorrect size, not accounting for the
alignment added to allocation. This leads to a memory leak.

It's possible MallocAllocatory still works fine, but the memory accounting
of the buffer is off. This leads to LeakCheckAllocator in BEF tests failing.

PiperOrigin-RevId: 341896455