[TFRT:Compiler] Add the tfrt-remove-device-attribute pass to remove the device attribute from corert.executeop.
Prior to this CL, the tf-to-corert pass will remove the device attribute from TF ope when converting TF op to a corert.executeop. After this pass, tf-to-corert will keep the device attribute in the corert.executeop, and add a tfrt-remove-device-attribute pass to the tf-executor-to-corert-pipeline pipeline. So the corert.executeop will still not have the device attribute after this pipeline.

The benefit of this approach is that we can run conversion pass after the tf-to-corert pass that can use the device attribute to determine whether we need to transfer values between devices.

PiperOrigin-RevId: 343771207