Add garbage collect and keep alive mechanisms to distributed TFRT.
Distributed contexts created by remote clients are subject to leaks if the remote clients are disconnected. To avoid it, we add garbage collect mechanism so that the servers will periodically check and delete stale contexts if they have been inactive for a certain period of time. The GC timeout is by default 600 seconds and configurable in ServerContextConfiguration.

The last access time for a distributed context is updated automatically by any RPCs with the context_id. The client also sends out KeepAlive messages periodically to make sure the remote contexts are alive. The time interval is half of the GC timeout (i.e., by default, 300 seconds).

PiperOrigin-RevId: 345709415