Add a utility function to make invoking a SyncBEFFunction easy, similar to invoking a C++ function.
Example:

Suppose a tfrt::Function foo takes two integer arguments and returns a float
value, we can invoke this function as follows:

  ExecutionContext exec_ctx = ...;  // Create an ExecutionContext.
  Expected<float> result = InvokeFunction<float>(foo, exec_ctx, int1, int2);

  if (!result) {
     // an error occurred.
     std::cout << "Error: " << result.takeError();
  } else {
    float result_value = *result;
  }

Similarly, for functions that have two return values, we can invoke the
function as follows.
  Expected<std::tuple<float, float>> result =
               InvokeFunction<float, float>(foo, exec_ctx, int1, int2);

If the function foo has no return value, we can invoke it as follows.
  Error error = InvokeFunction(foo, exec_ctx, int1, int2);

PiperOrigin-RevId: 345782915