Support AllGather of tensors with variable shapes
Unlike TF2, the work of finding out shapes of all tensors is moved to the kernel. The kernel does an extra round of AllGather to obtain variable shapes of participating tensors. This is needed because a tensor is copied into the output tensor at a particular order. The ordering depends on an axis and shapes of participating tensors.

The kernel also allocates the output tensor because the shape of output tensor is known after the extra round of AllGather-ing the shapes.

PiperOrigin-RevId: 353733788