Switch the positions of Functions and FunctionIndex sections in BEF
In BEF, the content of FunctionIndex section is necessary to decode
the content of the Functions section, but the FunctionIndex section
comes after the Functions section.

The positions have been switched to make it more streamed-read friendly.

Note that this change also swaps the section IDs of Functions and
FunctionIndex sections. It will prevent loading a BEF file created
before this change.

PiperOrigin-RevId: 355504840