Change BEFReader to return false for a failure case
BEFReader class functions were returning false for a success
and true for an error, and it was confusing.

Modifed the BEFReader class and its child class (BEFFileReader)
to return true for a success and false for an error.

Also renamed the member function ReadInt() to ReadVbrInt(), and
removed an unused member function ReadInt8().

PiperOrigin-RevId: 363318818