Make GPU resources hold a reference to the context.
We previously got away without it because the tfrt_cuda_test context was a primary context that lives indefinitely. If the context is non-primary, it is destroyed when no longer referenced, potentially before other resources (modules, streams, events, allocators) which are tied to the context are destroyed. The result of that is catastrophic: calling the CUDA API with a resource that has been destroyed generates a segmentation fault.

This change makes those resources hold a reference to the context. The overhead is not any larger than an explicit op to delete the resource.

This change also correctly ties modules to contexts instead of devices. It's a preliminary fix that requires further cleanup in a follow-up change (remove MultiDeviceModuleTable and tests).

PiperOrigin-RevId: 368180032