Introduce HostContextPool to manage HostContextPtr and fix the limitation of using circular buffer for allocating HostContexts.
In this CL, we replaced atomic with mutex for managing the HostContextPtr
allocations in order to fix the limitation of the circular buffer where an
allocation may fail even if there are less than kCapacity live HostContext
instances. Since HostContext allocation and deallocation are infrequent and are
not expected to be on the hot paths, the performance disadvantage of using
mutex as compared to atomic int does not matter here.

PiperOrigin-RevId: 368299333