Introduce a new AsyncValue semantic that keeps the payload value on error.
Normally, if the payload of an AsyncValue, if constructed, will be destructed
when the AsyncValue goes to the error state. This CL introduces a new semantic
that keeps the AsyncValue payload on error when the payload type inherits from
the `KeepAsyncValuePayloadOnError` struct. The new keep-payload-on-error semantic
is only intended to be used for the TpuRuntime. Thus, we keep the struct
`KeepAsyncValuePayloadOnError` in the tfrt::internal namespace to indicate this
restriction. Ideally, we should refactor the TpuRuntime code not to rely on
this semantic to keep the concept simpler and the usage of AsyncValue less
error-prone for TpuRuntime.

PiperOrigin-RevId: 369105769