Add GPU Platform and DnnDataType attributes. - `GetDeviceOp::platform()` returns a `wrapper::Platform`. - `DnnCreateTensorDescriptorOp::data_type()` returns a `wrapper::DnnDataType` (which is a discriminated union over `cudnnDataType_t` and `miopenDataType_t`).
Change tfrt_gpu.dnn.create_tensor_descriptor to take the data_type as attribute and drop the context argument. Instead of

```
  %data_type = tfrt.constant.ui32 0
  %desc = tfrt_gpu.dnn.create_tensor_descriptor %context, %data_type, %dims, %strides, %ch0
```
it is now
```
  %desc = tfrt_gpu.dnn.create_tensor_descriptor CUDNN_DATA_FLOAT, %dims, %strides, %ch0
```

I.e. it uses the data_type attribute to determine that we want to create a cuDNN tensor descriptor.

The attributes are I32Attr under the hood because BEF only encodes a fixed set of types. The kernels reconstruct the attribute values from the opaque `int`.

PiperOrigin-RevId: 371271088