[tfrt:jit] Add ConversionContext to ReturnValueConverter
ConversionContext allows to capture kernel specific operands information that can be passed back to the result conversion functions. Because throughout the cpurt we pass callbacks as llvm::function_ref to avoid copies context propagation is done in a very unsafe way, and it is the caller responsibility to extend the lifetime when needed.

PiperOrigin-RevId: 373661129