Do not copy BEF attribute into OpAttrs class
OpAttrsRawEntry is designed to hold an Op attribute entry within OpAttrs class.
An attribute payload is copied into OpAttrs class, but it is sub-optimal when
an attribute comes from a BEF attribute section.
Because BEF content is guaranteed to be available during execution, it is not needed
to copy attribute data into OpAttrs class.
Instead, we could simply keep a pointer.

PiperOrigin-RevId: 374462364