Move the two cast patterns from ConvertGpuToTfrtGpuPass to the pass that reconciles casts, so we don't need to make the cast op dynamically legal in the former pass.
This requires that ConvertCastToEventRecordPattern roots on the `chain, stream -> token` cast and rewrites the users (see explanation in `UnrealizedConversionCastPassthrough`). This has the advantage that we only create one event for multiple users.

Remove automatic `index -> ui32` conversion and perform it manually in the ConverLaunchFuncPattern instead.

Fixes `CastTo/FromChainAnd*()` names, which were the wrong way around.

PiperOrigin-RevId: 406223524