No longer return a chain from gpu module and global getter functions.
These functions previously returned a chain to signal that stream synchronization has completed. If no synchronization was needed, it would not return a chain. This required handling both cases in several places, and was especially awkward for global getters where the buffer could not be used until the chain signaled.

This change introduces an `%alias = tfrt_gpu.alias %value, %chain` op that allows the getter functions to return the `%value` only after the `%chain` has signaled. One can think of this op as a generalization of `tfrt.merge.chains`. The op is not specific to tfrt_gpu, but it's in that dialect because it's somewhat of an odd creature and we should use it with care.

Alternatives considered:
- Fix the case where `get_global` calls `get_module` which does not return a chain (either by returning a new chain, or by not returning a chain either). Instead of making the rewrite code simpler, it would have made it more complex.
- Change getters to always return a chain. This would introduce unnecessary runtime overhead and also make the IR program more complex.
- Use a different type for not-yet-initialized buffers. This would require tfrt_gpu.mem.copy to support this type as well.

This change also fixes the `tfrt.once` implementation to allow nested `tfrt.once` calls. Previously, it tried to re-lock the resource context mutex.

PiperOrigin-RevId: 412200402