Print source along with errors in executor test by passing locations through tfrt_gpu_opt.
`-mlir-print-debuginfo` adds locs to the output, so that bef references the test input instead of stdin and the diagnostic handler can refer to it. Add a test for that.

Remove cargo cult: don't add bef file to SourceMgr, it's not a source file.

Quiesce in tfrt_gpu_executor instead of just waiting for result.

PiperOrigin-RevId: 413063300