Hold on to a ref-count of `tfrt_gpu.mem.copy` operands until the async memcpy completes.
Before this change, operands without further use were potentially destroyed before the memcpy completed.

This is implemented by recording an event into the stream after the memcpy and holding on to the ref-counts until that event is signaled. The event is queried every 100ms as well as after a stream or event is synchronized.

TFRT had a low-latency event manager which got removed in cl/373100652. This change implements the suggested latency-tolerant approach instead.

PiperOrigin-RevId: 430997048