Write FFT wrapper to abstract over cuFFT and hipFFT
The FFT wrapper library to wrap both the cuFFT and hipFFT APIs. This wrapper is now low-level enough that we shouldn't need to break this abstraction layer when we implement TFRT performance improvements (such as plan reuse), but high-level enough to use e.g. TFRT's ownership model for FftHandles.

Implementing this requires adding support for FFT handle and plan creation, and well as a representation for FFT types.

We include FFT types based on the low-level (FFTW-style) API (such as D2Z),
instead of the high-level (XLA HLO) fft types (such as IRFFT).

PiperOrigin-RevId: 431575757