Add a work queue tailored to the requirements of gpu execution.
Tasks are executed in calling thread, except for blocking tasks which are run in an unbounded thread pool.

Schedule checking back on 'eventual callbacks' as blocking task (because it should not run in the calling thread) and remove callback from memcpy op (because it was deemed unnecessary and the default single-threaded work queue does not support blocking tasks).

Compared to the multi-threaded work queue that we have to use now (because the single-threaded work queue doesn't support blocking tasks used in collective ops), this reduces the MNIST iteration time from 191µs to 79µs.

PiperOrigin-RevId: 439915520