[JITRT] Allow to enable pass timing independently of debug output.
When enabling debug output, there is potentially a huge amount of IR printed
which slows down the compilation time. Therefore, it is helpful to have a
separate environment variable to enable pass timing. We will still enable pass
timing if the DEBUG_JITRT environment variable is set.

Also fix a bug introduced when trying to fix a use after free. The use after
free was fixed, but we need to capture the TimingScope after enabling timing.

PiperOrigin-RevId: 446386527