[tfrt] Replace AsyncValue error with an absl::Status
Mostly NFC, the only difference is that previously in theory AsyncValue error could carry the TFRT-specific decoded location information and now it can't.

However if location information is important the callers should use EmitError API and funnel it to the diagnostic handler. In the current code base we do not have any users of location information attached to async value errors.

PiperOrigin-RevId: 473492640