[tfrt] Alias TSL concurrency libraries from TFRT
This is a first step towards migrating to them, a process that we
will complete once we move the concurrency libraries to the TSL repo.
Until then, alias the TSL types in TFRT, which allows us to keep
the existing users unmodified without having two copies of the library,
and also migrate users who need it now (e.g. XLA Runtime) sooner.

Related: tensorflow/tensorflow#57844
PiperOrigin-RevId: 484602399