Prepare CUDA logging test for newer versions of cuDNN
With cuDNN 8.9.1 NVIDIA improved the error reporting in the library
(see https://docs.nvidia.com/deeplearning/cudnn/release-notes/index.html#rel-891 for details).

This unfortunately breaks the CUDA logging test that we have in TFRT since it changes the error message. So this CL
changes the test such that it accepts both the new and the old log message.

PiperOrigin-RevId: 552386787